+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/ir/transformers/CreateGraphTransformer.kt
@@ -3,21 +3,30 @@
 package dev.zacsweers.metro.compiler.ir.transformers
 
 import dev.zacsweers.metro.compiler.Symbols
+import dev.zacsweers.metro.compiler.expectAsOrNull
 import dev.zacsweers.metro.compiler.ir.IrMetroContext
 import dev.zacsweers.metro.compiler.ir.createIrBuilder
 import dev.zacsweers.metro.compiler.ir.implements
+import dev.zacsweers.metro.compiler.ir.irInvoke
+import dev.zacsweers.metro.compiler.ir.metroGraphOrFail
 import dev.zacsweers.metro.compiler.ir.rawType
+import dev.zacsweers.metro.compiler.ir.requireSimpleFunction
 import org.jetbrains.kotlin.ir.builders.irCall
+import org.jetbrains.kotlin.ir.builders.irCallConstructor
 import org.jetbrains.kotlin.ir.builders.irGetObject
 import org.jetbrains.kotlin.ir.expressions.IrCall
 import org.jetbrains.kotlin.ir.expressions.IrExpression
+import org.jetbrains.kotlin.ir.types.IrSimpleType
+import org.jetbrains.kotlin.ir.types.typeOrFail
 import org.jetbrains.kotlin.ir.util.classIdOrFail
 import org.jetbrains.kotlin.ir.util.companionObject
 import org.jetbrains.kotlin.ir.util.functions
 import org.jetbrains.kotlin.ir.util.hasAnnotation
+import org.jetbrains.kotlin.ir.util.isObject
 import org.jetbrains.kotlin.ir.util.kotlinFqName
 import org.jetbrains.kotlin.ir.util.nestedClasses
 import org.jetbrains.kotlin.ir.util.parentAsClass
+import org.jetbrains.kotlin.ir.util.primaryConstructor
 
 /**
  * Covers replacing createGraphFactory() compiler intrinsics with calls to the real graph factory.
@@ -39,10 +48,22 @@ internal object CreateGraphTransformer {
           val parentDeclaration = rawType.parentAsClass
           val companion = parentDeclaration.companionObject()!!
 
-          // If there's no $$Impl class, the companion object is the impl
-          val companionIsTheFactory =
-            companion.implements(rawType.classIdOrFail) &&
-              rawType.nestedClasses.singleOrNull { it.name == Symbols.Names.MetroImpl } == null
+          val factoryImpl = rawType.nestedClasses.find { it.name == Symbols.Names.MetroImpl }
+          if (factoryImpl != null) {
+            // Replace it with a call directly to the factory creator
+            return metroContext.createIrBuilder(expression.symbol).run {
+              if (factoryImpl.isObject) {
+                irGetObject(factoryImpl.symbol)
+              } else {
+                irInvoke(
+                  callee = companion.requireSimpleFunction(Symbols.StringNames.FACTORY),
+                  typeArgs = type.expectAsOrNull<IrSimpleType>()?.arguments?.map { it.typeOrFail },
+                )
+              }
+            }
+          }
+
+          val companionIsTheFactory = companion.implements(rawType.classIdOrFail)
 
           if (companionIsTheFactory) {
             return metroContext.createIrBuilder(expression.symbol).run {
@@ -75,6 +96,18 @@ internal object CreateGraphTransformer {
           // Already checked in FIR
           val rawType = type.rawType()
           val companion = rawType.companionObject()!!
+
+          val companionIsTheGraph = companion.implements(rawType.classIdOrFail)
+          if (companionIsTheGraph) {
+            return metroContext.createIrBuilder(expression.symbol).run {
+              val metroGraph = rawType.metroGraphOrFail
+              irCallConstructor(
+                metroGraph.primaryConstructor!!.symbol,
+                type.expectAsOrNull<IrSimpleType>()?.arguments.orEmpty().map { it.typeOrFail },
+              )
+            }
+          }
+
           val factoryFunction =
             companion.functions.singleOrNull {
               it.hasAnnotation(Symbols.FqNames.GraphFactoryInvokeFunctionMarkerClass)