+++ CHANGELOG.md
@@ -8,10 +8,12 @@ Changelog
     - support full IC compatibility with changing annotations and return types on `@Binds` and `@Multibinds` declarations
     - allow these declarations to be `private`
 - **Enhancement:** Allow `@Binds` and `@Multibinds` functions to be private.
+- **Enhancement:** Allow "static graphs" via companions implementing the graph interface itself.
 - **Fix:** When recording IC lookups of overridable declarations, only record the original declaration and not fake overrides.
 - **Fix:** Record IC lookups to `@Multibinds` declarations.
 - **Fix:** Write `@Multibinds` information to metro metadata.
 - **Fix:** Always write metro metadata to `@BindingContainer` classes, even if empty.
+- **Fix:** When `@Includes`-ing other graphs, link against the original interface accessor rather than the generated `$$MetroGraph` accessor.
 - **Fix:** Disambiguate contributed nullable bindings from non-nullable bindings.
 - Add a `ViewModel` assisted injection example to `compose-navigation-app` sample.
 - Small improvements to the doc site (404 page, favicon, etc.)